requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.31
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
stripe==7.8.0
jinja2==3.1.2
python-dateutil==2.8.2
psycopg2-binary==2.9.9
requests==2.31.0

